[git_tools] Make get_dev_null available from git_tools.
This is necessary to use get_dev_null from other places when using the
git function.
